requirements.txt
Django>=4.2
djangorestframework
django-environ
cx_Oracle
openai
gunicorn
uvicorn
